BS-13982 Redirection to the Tomcat homepage on IE 9,
process-autogenerated-form and step-autogenerated-form set $locationProvider.html5Mode to true,

but
IE9 do not support HTML5 History API. That is what angular html5Mode uses.
If we need our page run on IE9 we need to switch back or add code to detect if History API is available the use is or default to the hash
function supports_history_api()
{ return !!(window.history && history.pushState); }

Or stop enable html5Mode because we do not use it for the moment.