fix(tasklist): avoid errors due to Document object not being threadsafe when parsing forms.xml
Sometimes when several threads parse the same forms.xml, we get a NPE
while evaluating Xpath expressions. This happens for exemple when
running the selenium tests with the process templates.
This fix places the Xpath evaluation in a synchronized bloc to avoid
this issue.

covers [BS-14640](https://bonitasoft.atlassian.net/browse/BS-14640)