fix(looknfeel): add compiled bootstrap.css to less archive
In contrary to bonita.css, bootstrap.min.css is not compiled by portal while getting looknfeel from database.
That is why compiled bootstrap.min.css needs to be enbed in less archive.

Please also note that this is the less archive that is been used as portal looknfeel instead of css archive.