fix(BS-15313) handle the case were the caseID is not sent in the postMessage
With UID process start forms designed before 7.3, if you don't re-import
them into the UID, the case Id is not necessarily set in the post
Message sent by the pbButton. As a result, the portal is unable to
redirect to the details view of the case.
This commit allows to redirect to the home page of the current profile
instead of doing nothing (and potentially having the portal displayed
inside the portal if the targetURL on success of the form was set to
/bonita)

Covers [BS-15313](https://bonitasoft.atlassian.net/browse/BS-15313)