fix(CSRF API token) Clean API token value when invalidating (#207)
when a cookie already exists on a different path than the one expected,
we need to invalidate it. And Since there is no way of knowing the path
as it is not sent server-side (getPath return null) we can only
invalidate any cookie found. In addition this commit clean the cookie
value to avoid misunderstanding the response header.

Covers [BS-16241](https://bonitasoft.atlassian.net/browse/BS-16241)