feat(login): make login page responsive
Make login page responsive in a mobile first approach. CSS is for mobile and some rules are overriden for desktop via media query
This is not perfect for now since there are some overlapping issue on small screen and landscape.
Anyway, I believe this is way better than it was before.
Desktop display is unchanged